+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.0.69 (unreleased)
 
+* [Improve EnumField error messages to show enum members instead of values](https://github.com/anna-money/marshmallow-recipe/pull/227)
 * [Add JsonRawField with JSON validation for Any type fields](https://github.com/anna-money/marshmallow-recipe/pull/212)
 * [Fix UUID invalid_error not being applied](https://github.com/anna-money/marshmallow-recipe/pull/213)
 * [Fix int values being converted to float when dumping float fields](https://github.com/anna-money/marshmallow-recipe/pull/214)